# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lumen Copilot API
  version: 0.8.0
  description: >
    Source-of-truth contract for the Lumen Copilot REST surface. The backend
    implements to match this; the frontend client is generated from it.


    Revision 0.1.0 froze the **grounded-chat MVP wire** (issue #37,
    ADR-0006 Phase 0): auth, collections, documents, chat, and the model
    picker. Streaming assistant answers ride the WebSocket envelope
    (`websocket-envelopes.schema.json`); request/response stays REST (ADR-0003).


    This revision (0.2.0) **additively** freezes the **M2 Trust-Surfaces wire**
    (issue #80, ADR-0006 Phase 0 / ADR-0008 wave 0): `GET /search`
    (permission-trimmed, cited grounded search), `GET /audit` (the append-only
    audit trail with provenance), and the read-only `/admin/*` governance
    surfaces (members, model governance, risk tiers). These are the frozen
    hand-off artifact that gates the parallel M2 build slices; no shape from
    0.1.0 changed. All three are REST request/response — no new WebSocket
    envelope shapes were required.


    This revision (0.3.0) **additively** freezes the **Sources wire** (issue
    #108, ADR-0009 §5, ADR-0006 Phase 0): `GET /sources` (the connector grid),
    `POST /sources` (add a source — the first connector is a Web URL; an invalid
    or SSRF-blocked URL is a **422**, ADR-0009 §3), `POST /sources/{id}/sync`
    (re-sync), and `DELETE /sources/{id}` (remove, cascading the source's
...


...


...


...


...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /documents:
    get:
      operationId: listDocuments
      summary: List documents, optionally filtered by collection / status / text.
      tags: [documents]
      parameters:
        - name: collection_id
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Restrict to one collection.
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/DocumentStatus" }
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: Filter by filename substring (lexical, not semantic).
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: A page of documents.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: uploadDocumentLegacy
      summary: Retired binary upload endpoint.
      deprecated: true
      description: >
        Binary uploads through FastAPI are no longer accepted (spec 0008).
        Initiate a metadata-only session with `POST /api/v2/document-uploads`, upload
        each signed part directly to storage, then complete the session.
      tags: [documents]
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "410": { $ref: "#/components/responses/Gone" }
...
  /documents/{documentId}:
    parameters:
      - { $ref: "#/components/parameters/DocumentId" }
    get:
      operationId: getDocument
      summary: Get document metadata (incl. ingestion status).
      tags: [documents]
      responses:
        "200":
          description: The document.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteDocument
      summary: Delete a document and its chunks/embeddings.
      tags: [documents]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /documents/{documentId}/content:
    parameters:
      - { $ref: "#/components/parameters/DocumentId" }
    get:
      operationId: getDocumentContentLegacy
      summary: Retired backend byte/redirect endpoint.
      deprecated: true
      description: >
        Backend-proxied bytes and protected redirects are retired. Request a
        JSON capability from `/api/v2/documents/{documentId}/access-url`; the browser
        then loads/downloads directly from storage.
      tags: [documents]
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "410": { $ref: "#/components/responses/Gone" }
...
  /documents/{documentId}/text:
    parameters:
      - { $ref: "#/components/parameters/DocumentId" }
    get:
      operationId: getDocumentText
      summary: Get the extracted plain text of a ready document.
      description: >
        The ingestion parser's plain-text output for the document, reassembled
        from its stored chunks (issue #244). This is the viewer's text surface
        for formats a browser cannot render natively (DOCX/PPTX/XLSX) and the
        universal "show me the text" affordance. Visibility is identical to
        `/documents/{documentId}/content` (INV-2: not visible → 404, never 403).
        A document that exists but is not `ready` has no text yet → 409
        (`code=document_not_ready`).
      tags: [documents]
      responses:
        "200":
          description: The extracted text.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentText" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }

  # --- Artifacts (agent/run-produced files) : CC-B #208 / panel #222 ---------
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Opaque pagination cursor from a previous response's next_cursor.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Max items to return.
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    DocumentStatus:
      type: string
      description: Ingestion lifecycle.
      enum: [pending, processing, ready, failed]
    SourceLocation:
      type: object
      additionalProperties: false
      required: [kind, name, number, char_start, char_end]
      description: Exact half-open source-part span into retained extracted text. Backend guarantees 0 <= char_start <= char_end <= source length. An empty span denotes a blank part. The existing document text endpoint may return a truncated prefix of that source.
      properties:
        kind: { type: string, enum: [page, slide, sheet] }
        name: { type: string }
        number: { type: integer, minimum: 1 }
        char_start: { type: integer, minimum: 0 }
        char_end: { type: integer, minimum: 0 }
    ExtractionDiagnostics:
      type: object
      additionalProperties: false
      required: [character_count, replacement_characters, suspicious_controls, blank_parts, table_probe, warnings]
      description: Most recent native extraction attempt measurements and qualified text-presence heuristics; not proof of activated chunks or complete visual/table fidelity. Unknown coverage fields may be absent or null. Administrator-only within existing document visibility.
      properties:
        character_count: { type: integer, minimum: 0 }
        replacement_characters: { type: integer, minimum: 0 }
        suspicious_controls: { type: integer, minimum: 0 }
        source_part_kind: { type: [string, "null"], enum: [page, slide, sheet, null] }
        total_parts: { type: [integer, "null"], minimum: 0 }
        parts_with_text: { type: [integer, "null"], minimum: 0 }
        blank_parts: { type: array, items: { type: integer, minimum: 1 } }
        table_probe: { type: string, enum: [native_tables, sheet_cells, unavailable] }
        table_regions: { type: [integer, "null"], minimum: 0 }
        table_cells: { type: [integer, "null"], minimum: 0 }
        missing_table_cells: { type: [integer, "null"], minimum: 0 }
        warnings: { type: array, items: { type: string } }
    DocumentKind:
      type: string
      description: Processing/viewer family derived from the validated MIME type.
      enum: [document, audio, video]
    Document:
      type: object
      additionalProperties: false
      required:
        [id, filename, mime_type, size_bytes, collection_id, owner_id, kind,
         duration_ms, status, chunk_count, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        filename: { type: string }
        mime_type: { type: string }
        size_bytes: { type: integer, minimum: 0 }
        collection_id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/DocumentKind" }
        duration_ms:
          type: [integer, "null"]
          minimum: 0
          description: Player-relative media duration; null for ordinary documents.
        status: { $ref: "#/components/schemas/DocumentStatus" }
        source_locations:
          type: array
          description: Permissioned source-part map. Empty or absent means unknown until re-ingestion; never inferred for legacy documents.
          items: { $ref: "#/components/schemas/SourceLocation" }
        extraction_diagnostics:
          oneOf:
            - { $ref: "#/components/schemas/ExtractionDiagnostics" }
            - { type: "null" }
          description: Null for members or unknown legacy metadata; only visible administrators receive diagnostics. Administrator role does not widen document permissions.
        error:
          type: [string, "null"]
          description: Failure reason when status is failed.
        chunk_count:
          type: integer
          minimum: 0
          description: Number of indexed chunks (0 until ingestion completes).
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DocumentList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Document" }
        next_cursor:
          type: [string, "null"]
    DocumentText:
      type: object
      additionalProperties: false
      required: [text, chunk_count, truncated]
      properties:
        text:
          type: string
          description: The retained extracted plain text; legacy documents fall back to chunk reconstruction.
        chunk_count:
          type: integer
          minimum: 0
          description: Number of chunks the text was reassembled from.
        truncated:
          type: boolean
          description: >
            True when the text was capped at the server limit
            (DOCUMENT_TEXT_MAX_BYTES); the UI should say so and offer the
            original-bytes download instead of implying completeness.
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: "about:blank"
          description: A URI identifying the problem type.
        title: { type: string, description: Short, human-readable summary. }
        status: { type: integer, description: HTTP status code. }
        detail: { type: string, description: Human-readable explanation for this occurrence. }
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Stable, machine-readable error code for clients to branch on.
        errors:
          type: array
          description: Field-level validation errors, when applicable.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }

    # --- User preferences (spec 0005) ---------------------------------------
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, or expired credentials (INV-4).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: Not found — or in another tenant / not permitted (INV-1/INV-2).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: >
        The resource exists but is not in a state that allows this request
        (illegal transition / not ready — INV-8). The problem `code` names the
        specific conflict (e.g. `document_not_ready`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Gone:
      description: >
        This operation was deliberately retired. The problem `code` names the
        replacement path; the server does not parse or proxy a legacy file body.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  securitySchemes:
    # Identity/tenant is resolved in backend/app/auth (CC-3, spec 0004 §2.3):
    # app-managed email+password → short-lived access JWT. Bearer-token on every
    # endpoint except health + login + refresh.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...